Move commands to alluxio.cli package to loadCommand by CommandUtil
