+++ integration/fuse/src/main/java/alluxio/cli/metadatacache/DropAllCommand.java
@@ -9,7 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.fuse.cli.metadatacache;
+package alluxio.cli.metadatacache;
 
 import alluxio.AlluxioURI;
 import alluxio.Constants;
@@ -19,43 +19,38 @@ import alluxio.client.file.URIStatus;
 import alluxio.conf.AlluxioConfiguration;
 import alluxio.conf.PropertyKey;
 import alluxio.exception.status.InvalidArgumentException;
-import alluxio.fuse.cli.command.AbstractFuseShellCommand;
+import alluxio.cli.command.AbstractFuseShellCommand;
 import alluxio.wire.FileInfo;
 
-public final class SizeCommand extends AbstractFuseShellCommand {
-  public SizeCommand(FileSystem fs, AlluxioConfiguration conf, String parentCommandName) {
+public final class DropAllCommand extends AbstractFuseShellCommand {
+  public DropAllCommand(FileSystem fs, AlluxioConfiguration conf, String parentCommandName) {
     super(fs, conf, parentCommandName);
   }
 
   @Override
   public String getCommandName() {
-    return "size";
+    return "dropAll";
   }
 
   @Override
   public String getUsage() {
-    return String.format("ls -l %s%s.%s.%s", Constants.DEAFULT_FUSE_MOUNT,
+    return String.format("%s%s.%s.%s", Constants.DEAFULT_FUSE_MOUNT,
         Constants.ALLUXIO_CLI_PATH, getParentCommandName(), getCommandName());
   }
 
   @Override
-  public URIStatus run(AlluxioURI path, String [] argv) {
-    // The 'ls -l' command will show metadata cache size in the <filesize> field.
-    long size = ((MetadataCachingBaseFileSystem) mFileSystem).getMetadataCacheSize();
-    return new URIStatus(new FileInfo().setLength(size).setCompleted(true));
-  }
-
-  @Override
-  public String getDescription() {
-    return "Get fuse client metadata size.";
-  }
-
-  @Override
-  public void validateArgs(String[] argv) throws InvalidArgumentException {
+  public URIStatus run(AlluxioURI path, String [] argv) throws InvalidArgumentException {
     if (!mConf.getBoolean(PropertyKey.USER_METADATA_CACHE_ENABLED)) {
       throw new UnsupportedOperationException(String.format("metadatacache command is "
               + "not supported when %s is false",
           PropertyKey.USER_METADATA_CACHE_ENABLED.getName()));
     }
+    ((MetadataCachingBaseFileSystem) mFileSystem).dropMetadataCacheAll();
+    return new URIStatus(new FileInfo().setCompleted(true));
+  }
+
+  @Override
+  public String getDescription() {
+    return "Clear all the fuse client metadata cache.";
   }
 }
