+++ integration/fuse/src/main/java/alluxio/cli/metadatacache/DropCommand.java
@@ -9,17 +9,17 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.fuse.cli.metadatacache;
+package alluxio.cli.metadatacache;
 
 import alluxio.AlluxioURI;
 import alluxio.Constants;
+import alluxio.cli.command.AbstractFuseShellCommand;
 import alluxio.client.file.FileSystem;
 import alluxio.client.file.MetadataCachingBaseFileSystem;
 import alluxio.client.file.URIStatus;
 import alluxio.conf.AlluxioConfiguration;
 import alluxio.conf.PropertyKey;
 import alluxio.exception.status.InvalidArgumentException;
-import alluxio.fuse.cli.command.AbstractFuseShellCommand;
 import alluxio.wire.FileInfo;
 
 public final class DropCommand extends AbstractFuseShellCommand {
@@ -34,13 +34,18 @@ public final class DropCommand extends AbstractFuseShellCommand {
 
   @Override
   public String getUsage() {
-    return String.format("ls -l %s%s%s.%s.%s", Constants.DEAFULT_FUSE_MOUNT,
+    return String.format("%s%s%s.%s.%s", Constants.DEAFULT_FUSE_MOUNT,
         "/path/to/clean", Constants.ALLUXIO_CLI_PATH, getParentCommandName(),
         getCommandName());
   }
 
   @Override
-  public URIStatus run(AlluxioURI path, String [] argv) {
+  public URIStatus run(AlluxioURI path, String [] argv) throws InvalidArgumentException {
+    if (!mConf.getBoolean(PropertyKey.USER_METADATA_CACHE_ENABLED)) {
+      throw new UnsupportedOperationException(String.format("metadatacache command is "
+              + "not supported when %s is false",
+          PropertyKey.USER_METADATA_CACHE_ENABLED.getName()));
+    }
     ((MetadataCachingBaseFileSystem) mFileSystem).dropMetadataCache(path);
     return new URIStatus(new FileInfo().setCompleted(true));
   }
@@ -48,14 +53,5 @@ public final class DropCommand extends AbstractFuseShellCommand {
   @Override
   public String getDescription() {
     return "Clear the specific fuse path metadata cache.";
-  }
-
-  @Override
-  public void validateArgs(String[] argv) throws InvalidArgumentException {
-    if (!mConf.getBoolean(PropertyKey.USER_METADATA_CACHE_ENABLED)) {
-      throw new UnsupportedOperationException(String.format("metadatacache command is "
-              + "not supported when %s is false",
-          PropertyKey.USER_METADATA_CACHE_ENABLED.getName()));
-    }
   }
 }
